Use unix time stamp to create local id for offline tasks. Change project_id and related to int64.

+++ utasklist.pas
@@ -397,7 +397,7 @@ begin
       l.Font.Color := clGray;
       l.Font.Size := FONTSIZESMALL;
       l.Align := alLeft;
-      l.Caption := FPaymo.GetProjectName(arr[i].GetPath('project_id').AsInteger) +
+      l.Caption := FPaymo.GetProjectName(arr[i].GetPath('project_id').AsInt64) +
         '  [' + UTF8UpperCase(FormatDateTime('ddddd',
         StringToDateTime(arr[i].GetPath('created_on').AsString))) + ']';
       l.OnClick := @OnClickItemParent;
@@ -685,7 +685,7 @@ begin
       l.Font.Color := clGray;
       l.Font.Size := FONTSIZESMALL;
       l.Align := alLeft;
-      l.Caption := FPaymo.GetProjectName(arr[i].GetPath('project_id').AsInteger) +
+      l.Caption := FPaymo.GetProjectName(arr[i].GetPath('project_id').AsInt64) +
         '  [' + UTF8UpperCase(FormatDateTime('ddddd',
         StringToDateTime(arr[i].GetPath('created_on').AsString))) + ']';
       //l.OnClick := @OnClickItemParent;
